Bugfix: Correct paths to inventory and configuration file.
Even if this is already working and it seems to be SC3 independent, we need
to debug and improve it much more. An example: there are two instancies of
the inventory. One is loaded inthe application and another one is loaded by
the Routing Service. This needs to be corrected, the one loaded by the
application could be removed.